Added gates to force players to solve offices in order
